requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0
mysql-connector-python==9.3.0